Adding support for GlobalResults extraction by requests
