+++ GSA_Adapter/CRUD/ReadResults.cs
@@ -71,64 +71,6 @@ namespace BH.Adapter.GSA
 
         }
 
-        /***************************************************/
-
-        private IEnumerable<IResult> ExtractGlobalReaction(List<int> cases)
-        {
-            List<GlobalReactions> reactions = new List<GlobalReactions>();
-
-            string id = ""; //TODO: Strategy for ids for full model
-
-            foreach (int loadCase in cases)
-            {
-                List<string> forceStrings = new List<string>();
-                List<string> momentStrings = new List<string>();
-
-                forceStrings.Add(m_gsaCom.GwaCommand("GET, TOTAL_FORCE, " + loadCase + ",REACT"));
-                forceStrings.Add(m_gsaCom.GwaCommand("GET, TOTAL_FORCE, " + loadCase + ",SUPPORT"));
-                forceStrings.Add(m_gsaCom.GwaCommand("GET, TOTAL_FORCE, " + loadCase + ",SPRING"));
-                forceStrings.Add(m_gsaCom.GwaCommand("GET, TOTAL_FORCE, " + loadCase + ",SOIL"));
-
-                momentStrings.Add(m_gsaCom.GwaCommand("GET, TOTAL_MOMENT, " + loadCase + ",REACT"));
-                momentStrings.Add(m_gsaCom.GwaCommand("GET, TOTAL_MOMENT, " + loadCase + ",SUPPORT"));
-                momentStrings.Add(m_gsaCom.GwaCommand("GET, TOTAL_MOMENT, " + loadCase + ",SPRING"));
-                momentStrings.Add(m_gsaCom.GwaCommand("GET, TOTAL_MOMENT, " + loadCase + ",SOIL"));
-
-                //string force = m_gsaCom.GwaCommand("GET, TOTAL_FORCE, " + loadCase + ",REACT");
-                //string moment = m_gsaCom.GwaCommand("GET, TOTAL_MOMENT, " + loadCase + ",REACT");
-                //reactions.Add(BH.Engine.GSA.Convert.ToBHoMGlobalReactions(id, force, moment));
-
-                reactions.Add(BH.Engine.GSA.Convert.ToBHoMGlobalReactions(id, forceStrings, momentStrings));
-            }
-
-            return reactions;
-        }
-
-        /***************************************************/
-
-        private IEnumerable<IResult> ExtractGlobalDynamics(List<int> cases)
-        {
-            List<ModalDynamics> dynamics = new List<ModalDynamics>();
-
-            string id = ""; //TODO: Strategy for ids for full model
-
-            foreach (int loadCase in cases)
-            {
-                string mode = m_gsaCom.GwaCommand("GET, MODE, " + loadCase);
-
-                if (String.IsNullOrWhiteSpace(mode))
-                    continue;
-                string frequency = m_gsaCom.GwaCommand("GET, FREQ, " + loadCase);
-                string mass = m_gsaCom.GwaCommand("GET, MASS, " + loadCase);
-                string damping = m_gsaCom.GwaCommand("GET, MODAL_DAMP, " + loadCase);
-                string stiffness = m_gsaCom.GwaCommand("GET, MODAL_STIFF, " + loadCase);
-                string effMassTran = m_gsaCom.GwaCommand("GET, EFF_MASS, " + loadCase + ",TRAN");
-                string effMassRot = m_gsaCom.GwaCommand("GET, EFF_MASS, " + loadCase + ",ROT");
-                dynamics.Add(BH.Engine.GSA.Convert.ToBHoMModalDynamics(id, mode, frequency, mass, stiffness, damping, effMassTran, effMassRot));
-            }
-
-            return dynamics;
-        }
 
         /***************************************************/
         /**** Object Results  Methods                   ****/
