+++ GSA_Adapter/GSA_Adapter.csproj
@@ -103,6 +103,7 @@
     <Compile Include="AdapterActions\Execute.cs" />
     <Compile Include="CRUD\ReadResults.cs" />
     <Compile Include="CRUD\ReadResultsElements.cs" />
+    <Compile Include="CRUD\ReadResultsGlobal.cs" />
     <Compile Include="CRUD\UpdateObjects.cs" />
     <Compile Include="CRUD\UpdateProperty.cs" />
     <Compile Include="Types\DependencyTypes.cs" />
